ref(structure): moving plugins into a separate dir
