--- js/deepLinker.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.DeepLinker = void 0;
-class DeepLinker {
-}
-exports.DeepLinker = DeepLinker;
