+++ js/plugins/deep-linking/DeepLinkingProvider.d.ts
@@ -1,3 +1,3 @@
-export declare abstract class DeepLinker {
+export declare abstract class DeepLinkingProvider {
     abstract subscribe(handler?: (token: string) => Promise<void>): void;
 }
