+++ js/plugins/deep-linking/DeepLinkingProvider.js
@@ -1,0 +1,6 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.DeepLinkingProvider = void 0;
+class DeepLinkingProvider {
+}
+exports.DeepLinkingProvider = DeepLinkingProvider;
