+++ js/plugins/web-sockets/JolocomWebSockets.d.ts
@@ -1,0 +1,4 @@
+import { JolocomSDK, JolocomPlugin } from '@jolocom/sdk';
+export declare class JolocomWebSockets implements JolocomPlugin {
+    register(sdk: JolocomSDK): Promise<void>;
+}
