+++ ts/@types/react-native-randombytes/index.d.ts
@@ -1,0 +1,3 @@
+declare module 'react-native-randombytes' {
+  export function randomBytes(length: number): Buffer
+}
