+++ ts/plugins/deep-linking/DeepLinkingProvider.ts
@@ -1,3 +1,3 @@
-export abstract class DeepLinker {
+export abstract class DeepLinkingProvider {
   public abstract subscribe(handler?: (token: string) => Promise<void>): void
 }
